PrismApplication simplification
