+++ Source/Windows10/Prism.Windows/StartArgs.cs
@@ -1,6 +1,5 @@
 ﻿using Windows.ApplicationModel.Activation;
 
-
 namespace Prism
 {
     public class StartArgs : IStartArgs
